part3/pyproject.toml
[tool.poetry]
name = "part3"
version = "0.1.0"
description = ""
authors = ["Metin100 <9455@holbertonstudents.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-restx = "^1.3.0"
flask-bcrypt = "^1.0.1"
flask-jwt-extended = "^4.7.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
mysqlclient = "^2.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
